fixes #290 ImageIndexViewをリファクタ
